decode_token ready for decoding from the Authorization header and using if we need the email address somewhere on the doc-index-updater call (ie audit)
