+++ medicines/doc-index-updater/Cargo.toml
@@ -17,6 +17,7 @@ bytes = "*"
 chrono = "0.4.11"
 futures = "0.3.4"
 hyper = "0.13"
+jsonwebtoken = "7.1.0"
 lazy_static = "1.4.0"
 md5 = "0.7.0"
 percent-encoding = "2.1.0"
@@ -38,7 +39,7 @@ tracing-subscriber = "0.2.3"
 url = "2.1.1"
 uuid = { version = "0.8", features = ["serde", "v4"] }
 warp = { git = "https://github.com/m-doughty/warp", branch = "add-xml-support" }
-jsonwebtoken = "7.1.0"
+
 
 [dev-dependencies]
 test-case = "1.0.0"
